✨ feat: create lodge types and refactor imports for better structure
