+++ src/types/lodge.ts
@@ -1,0 +1,59 @@
+export interface Lodge {
+  id: number;
+  name: string;
+  address: string;
+  latitude: number;
+  longitude: number;
+  description: string | null;
+  accommodationType: string;
+  roomTypes: RoomType[];
+  hotSpringLodgeImage: LodgeImage[];
+}
+
+export interface LodgeImage {
+  id: number;
+  lodgeId: number;
+  imageUrl: string;
+}
+
+export interface RoomType {
+  id?: number;
+  name: string;
+  description: string | null;
+  basePrice: number;
+  weekendPrice?: number;
+  maxAdults: number;
+  maxChildren: number;
+  totalRooms: number;
+  seasonalPricing?: SeasonalPricing[];
+  roomTypeImage: RoomTypeImage[];
+}
+
+export interface RoomTypeImage {
+  id: number;
+  roomTypeId: number;
+  imageUrl: string;
+}
+
+export interface SeasonalPricing {
+  from: string;
+  to: string;
+  basePrice: number;
+  weekendPrice: number;
+}
+
+export interface RoomInventory {
+  id: number;
+  lodgeId: number;
+  roomTypeId: number;
+  date: string;
+  availableRooms: number;
+}
+
+export interface RoomPricing {
+  id: number;
+  roomTypeId: number;
+  date: string;
+  price: number;
+  priceType: "WEEKDAY" | "WEEKEND" | "PEAK" | "OFF";
+}
